+++ Framework/Emlid.WindowsIoT.Hardware/Components/Mb85rcv/Mb85rc256vDevice.cs
@@ -38,7 +38,7 @@ namespace Emlid.WindowsIot.Hardware.Components.Mb85rcv
         public const byte Density = 0x05;
 
         /// <summary>
-        /// Maximum number of devices for this model.
+        /// Maximum number of devices (chip number) for this model.
         /// </summary>
         public const int MaximumDevices = 8;
 
@@ -104,7 +104,7 @@ namespace Emlid.WindowsIot.Hardware.Components.Mb85rcv
         /// </summary>
         /// <param name="controller">I2C controller on which the device is connected.</param>
         /// <param name="chipNumber">
-        /// Device (chip code, not FRAM memory) address from zero to the supported <see cref="MaximumDevices"/>.
+        /// Device (chip) number, from zero to the <see cref="MaximumDevices"/> supported.
         /// </param>
         /// <returns>Device ID.</returns>
         [CLSCompliant(false)]
@@ -127,10 +127,10 @@ namespace Emlid.WindowsIot.Hardware.Components.Mb85rcv
         /// Gets the I2C address for data commands with the specified chip number (device address code).
         /// </summary>
         /// <param name="chipNumber">
-        /// Device (chip code, not FRAM memory) address from zero to the supported <see cref="MaximumDevices"/>.
+        /// Device (chip) number, from zero to the <see cref="MaximumDevices"/> supported.
         /// </param>
         /// <returns>7-bit I2C address.</returns>
-        public static byte GetDataI2cAddress(int chipNumber)
+        public static byte GetDataI2cAddress(byte chipNumber)
         {
             // Validate
             if (chipNumber < 0 || chipNumber > MaximumDevices)
@@ -144,10 +144,10 @@ namespace Emlid.WindowsIot.Hardware.Components.Mb85rcv
         /// Gets the I2C address for the device ID command with the specified chip number (device address code).
         /// </summary>
         /// <param name="chipNumber">
-        /// Device (chip code, not FRAM memory) address from zero to the supported <see cref="MaximumDevices"/>.
+        /// Device (chip) number, from zero to the <see cref="MaximumDevices"/> supported.
         /// </param>
         /// <returns>7-bit I2C address.</returns>
-        public static byte GetDeviceIdI2cAddress(int chipNumber)
+        public static byte GetDeviceIdI2cAddress(byte chipNumber)
         {
             // Validate
             if (chipNumber < 0 || chipNumber > MaximumDevices)
